Changed api and added buttons to nodered
