+++ node-red/data/.config.users.json.backup
@@ -11,6 +11,9 @@
                 "view-node-show-label": true,
                 "view-show-tips": true,
                 "view-show-welcome-tours": true
+            },
+            "tours": {
+                "welcome": "4.0.8"
             }
         }
     }
