+++ node-red/data/package.json
@@ -5,6 +5,7 @@
     "private": true,
     "dependencies": {
         "node-red-contrib-postgresql": "~0.14.2",
+        "node-red-dashboard": "~3.6.5",
         "node-red-project": "file:"
     }
 }
